+++ packages/theme-validate/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./liquid.js";
 export * from "./json.js";
 export * from "./paths.js";
+export * from "./schema.js";
 export * from "./semver.js";
 export * from "./themeInfo.js";